[WIP] Combining search results middleware
