+++ webpack/store/configure_store.js
@@ -3,7 +3,6 @@ import promiseMiddleware from 'redux-promise-middleware';
 import createLogger from 'redux-logger';
 
 import questionsFromResponseSets from '../middleware/questions_from_response_sets';
-import questionsFromSearchResult from '../middleware/questions_from_search_result';
 import questionsFromForms from '../middleware/questions_from_forms';
 import responseSetsFromQuestions from '../middleware/response_sets_from_questions';
 import responseSetsFromForms from '../middleware/response_sets_from_forms';
@@ -14,7 +13,7 @@ import parentFromSurveys from '../middleware/parent_from_surveys';
 import responseTypesFromQuestions from '../middleware/response_types_from_questions';
 import questionTypesFromQuestions from '../middleware/question_types_from_questions';
 import stuffFromSurveys from '../middleware/stuff_from_surveys';
-import formsFromSearchResult from '../middleware/forms_from_search_result';
+import extractFromSearchResults from '../middleware/extract_from_search_results';
 import unauthenticatedResponse from '../middleware/unauthenticated_response';
 import responseSetsFromSearch from '../middleware/response_set_from_search';
 
@@ -27,7 +26,6 @@ export default function configureStore(initialState) {
     unauthenticatedResponse,
     questionsFromResponseSets,
     questionsFromForms,
-    questionsFromSearchResult,
     responseSetsFromQuestions,
     parentFromResponseSets,
     parentFromQuestions,
@@ -38,7 +36,7 @@ export default function configureStore(initialState) {
     responseSetsFromForms,
     responseSetsFromSearch,
     stuffFromSurveys,
-    formsFromSearchResult
+    extractFromSearchResults
   );
 
   // Sets up http://zalmoxisus.github.io/redux-devtools-extension/
